# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Sarthi Companion API
  version: 2.0.0
  description: |
    Canonical backend contract for the Sarthi companion rebuild.
    This contract replaces the older MVP resolution/week-1 APIs.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /v1/teams:
    get:
      tags: [teams]
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamListResponse"
    post:
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamCreateRequest"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamCreateResponse"
  /v1/teams/join:
    post:
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamJoinRequest"
      responses:
        "200":
          description: Joined team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetailResponse"
  /v1/teams/{team_id}:
    parameters:
      - $ref: "#/components/parameters/teamIdPath"
    get:
      tags: [teams]
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetailResponse"
    delete:
      tags: [teams]
      responses:
        "204":
          description: Deleted team
  /v1/teams/{team_id}/leave:
    post:
      tags: [teams]
      parameters:
        - $ref: "#/components/parameters/teamIdPath"
      responses:
        "204":
          description: Left team
...
  /v1/brain-dumps:
    get:
      tags: [brain-dumps]
      responses:
        "200":
          description: Brain dumps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainDumpListPage"
    post:
      tags: [brain-dumps]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrainDumpCreateRequest"
      responses:
        "201":
          description: Captured brain dump
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainDump"
  /v1/focus-sessions/{session_id}/brain-dumps:
    post:
      tags: [brain-dumps, focus-sessions]
      parameters:
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrainDumpCreateRequest"
      responses:
        "201":
          description: Captured focus brain dump
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainDump"
  /v1/brain-dumps/{dump_id}:
    get:
      tags: [brain-dumps]
      parameters:
        - $ref: "#/components/parameters/dumpId"
      responses:
        "200":
          description: Brain dump
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainDump"
  /v1/brain-dumps/{dump_id}/apply:
    post:
      tags: [brain-dumps]
      parameters:
        - $ref: "#/components/parameters/dumpId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrainDumpApplyRequest"
      responses:
        "200":
          description: Applied proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainDumpApplyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    teamIdPath:
      name: team_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    sessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    dumpId:
      name: dump_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    TeamCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    TeamJoinRequest:
      type: object
      required: [invite_code]
      properties:
        invite_code: { type: string }
    TeamCreateResponse:
      type: object
      additionalProperties: true
    TeamDetailResponse:
      type: object
      additionalProperties: true
    TeamListResponse:
      type: object
      additionalProperties: true
...
    BrainDumpCreateRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 20000 }
        focus_session_id: { type: [string, "null"], format: uuid }
        task_id: { type: [string, "null"], format: uuid }
        goal_id: { type: [string, "null"], format: uuid }
        team_id: { type: [string, "null"], format: uuid }
    BrainDump:
      type: object
      additionalProperties: true
    BrainDumpListPage:
      type: object
      additionalProperties: true
    BrainDumpApplyRequest:
      type: object
      properties:
        proposal_ids:
          type: array
          items: { type: string, format: uuid }
    BrainDumpApplyResponse:
      type: object
      additionalProperties: true
...